[#31] 피드백 반영
- FollowController URL 수정
- FollowService에서 알림 기능을 처리하도록 수정
- FollowService 인터페이스 제거
