+++ src/main/java/me/liiot/snsserver/controller/FollowController.java
@@ -3,9 +3,7 @@ package me.liiot.snsserver.controller;
 import lombok.RequiredArgsConstructor;
 import me.liiot.snsserver.annotation.CheckLogin;
 import me.liiot.snsserver.annotation.CurrentUser;
-import me.liiot.snsserver.enumeration.AlarmType;
 import me.liiot.snsserver.model.user.User;
-import me.liiot.snsserver.service.AlarmService;
 import me.liiot.snsserver.service.FollowService;
 import me.liiot.snsserver.util.HttpResponses;
 import org.springframework.http.ResponseEntity;
@@ -13,12 +11,10 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/follow")
+@RequestMapping("/follows")
 public class FollowController {
 
     private final FollowService followService;
-
-    private final AlarmService alarmService;
 
     @PostMapping("/{followUserId}")
     @CheckLogin
@@ -26,7 +22,6 @@ public class FollowController {
                                            @CurrentUser User currentUser) {
 
         followService.addFollowList(currentUser.getUserId(), followUserId);
-        alarmService.addAlarm(currentUser.getUserId(), followUserId, AlarmType.FOLLOWING);
 
         return HttpResponses.RESPONSE_CREATED;
     }
